adapter-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adapter-agent"
version = "1.0.0"
description = "AI安全工具适配部署Agent - 完整版（四层两体系架构）"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "GuardAdapter Team"}
]
keywords = ["ai-safety", "guardrails", "security", "deployment", "agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # CLI & UI
    "click>=8.0.0",
    "rich>=13.0.0",
    # API Framework
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    # Data & Config
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Async
    "aiohttp>=3.8.0",
    "httpx>=0.24.0",
    # Template & Code Gen
    "jinja2>=3.0.0",
    # Git & Project Analysis
    "gitpython>=3.1.0",
    # Vector DB & RAG
    "chromadb>=0.4.0",
    # LLM Integration
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    # Monitoring
    "prometheus-client>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
training = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "peft>=0.6.0",
    "trl>=0.7.0",
]

[project.scripts]
adapter-agent = "adapter_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["adapter_agent"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "UP"]